Bilal branch3

+++ investor/src/main/java/be/vinci/ipl/vsx/investor/InvestorService.java
@@ -1,27 +1,34 @@
 package be.vinci.ipl.vsx.investor;
 
-import jakarta.persistence.criteria.CriteriaBuilder;
+import be.vinci.ipl.vsx.investor.models.Credentials;
+import be.vinci.ipl.vsx.investor.models.Investor;
+import be.vinci.ipl.vsx.investor.models.InvestorWithPassword;
+import be.vinci.ipl.vsx.investor.repositories.AuthenticationProxy;
+import be.vinci.ipl.vsx.investor.repositories.InvestorRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.swing.*;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class InvestorService {
 
 
     private final InvestorRepository investorRepository;
+    private final AuthenticationProxy authenticationProxy;
 
 
     /**
      * Constructor for InvestorService.
      *
-     * @param investorRepo The repository for handling investor data.
+     * @param investorRepo        The repository for handling investor data.
+     * @param authenticationProxy
      */
     @Autowired
-    public InvestorService(InvestorRepository investorRepo){
+    public InvestorService(InvestorRepository investorRepo, AuthenticationProxy authenticationProxy){
         this.investorRepository = investorRepo;
+        this.authenticationProxy = authenticationProxy;
     }
 
     public List<Investor> readAll(){
@@ -45,11 +52,18 @@ public class InvestorService {
      * @param investor The investor data to be created.
      * @return The created investor, or null if the investor already exists.
      */
-    public Investor createInvestor(Investor investor) {
-        if(investorRepository.existsById(investor.getUsername())){
-            return null;
+    public boolean createInvestor(InvestorWithPassword  investor) {
+        if(investorRepository.existsById(investor.getInvestor().getUsername())){
+            return false;
         }
-        return investorRepository.save(investor);
+
+        Credentials cerdential = new Credentials();
+        cerdential.setUsername(investor.getInvestor().getUsername());
+        cerdential.setPassword(investor.getPassword());
+        authenticationProxy.createCredentials(cerdential.getUsername(),cerdential);
+
+        investorRepository.save(investor.getInvestor());
+        return true;
     }
 
 
@@ -71,7 +85,14 @@ public class InvestorService {
      *
      * @param username The username of the investor to be deleted.
      */
-    public void deleteInvestor(String username) {
+    public boolean deleteInvestor(String username) {
+        if (!investorRepository.existsById(username)){
+            return false;
+        }
+
+        authenticationProxy.deleteCredentials(username);
+
         investorRepository.deleteById(username);
+        return true;
     }
 }
